examples: bump and unify the tfjs versions
I'm not sure if there's a better way to do this. peer-dependencies
_seems_ like what we'd want to use here, but that also means (afaik)
that running `npm install` won't install tfjs and friends in client and
I'm not sure what that'll do to the bundling process.

I listed tfjs as a peer-dependency (with a specific pinned version) even
though it's also a dependency; maybe this'll at least warn users of the
client package when they're using an incompatible version of tfjs
upstream.

Also, the way npm resolves packages this _shouldn't be an issue_ but it
very much seems like it is. Maybe something about parcel's
tree-shaking/dedup process.