client: switch back to rollup!
third time, charm, etc., etc.

But actually, this is the same configuration that we had before
switching to parcel (so, it does indeed seem to be because of the
tensorflowjs version mismatch).

faster builds of examples/client + better tree shaking (significantly
smaller js bundles in dist/) are the upside to this