server: bump tensorflow version
The 2.0 betas didn't have the Delegate API so we're switching to using the nightly 2.0 preview.
This actually worked right out of the box; installing both the stable and the preview version was probably a bad idea and is _probably_ what required us to pin the subpackages to specific versions.

So, anyways, we can ditch the subpackages in the Pipfile now.