build: minor bugs
Something seems to cause the client folder to be modified just after a build in it finishes, so we'll track the src folder inside it instead. Not perfect, but it'll do for now.
Thanks to the addition of the mypy protoc plugin, the build script is now dependent on dev dependencies being installed.