client: run the tests on node instead of telling jest to try to emulate a browser
This is very much not optimal, but currently the tests do not run in jest with jsdom (even with jsdom 14 - "jest-environment-jsdom-fourteen"), because it's missing TextEncoder/TextDecoder.
There's an open issue [here](https://github.com/jsdom/jsdom/issues/2524) - until this is resolved, we're going to have to run as node.
The jest folks add TextEncoder and TextDecoder to the node environment (jest-environment-node) (see [here](https://github.com/facebook/jest/pull/8022)), so running as node works.

Tensorflow.js gives us a nice warning when running as node (about actually using the native backend) which we're also just going to have to live with.

It's also worth noting that these (TextEncoder, TextDecoder) aren't exactly new APIs:
  - https://developer.mozilla.org/en-US/docs/Web/API/TextDecoder#Browser_compatibility
  - https://caniuse.com/#feat=textencoder