server: first pass at ncore support
Creating/adding the delegate when its present + restricting the model store to 1 model when using ncore is really what this commit added.

This isn't complete until the model API exists since this only really works if clients are able to request that models be loaded (we'll load the first one and ignore the others).
Finish the model API and having it actually be useful probably means finishing conversion too.

Also, another way to handle models after the 1st one on NCore is to load them _without_ the delegate, but I don't want to do this since it'd be more things to communicate to the client and I think it could easily lead to confusion and hair pulling:
(why is this not running on NCore!? oh, someone refreshed another demo page right after the server restarted..)