client: expose the actual error to users
Here we're exposing the protobuf objects in an interface (which we've been careful not to do elsewhere), but this is probably fine. It's unlikely that we're going to do much with the errors we get other than print them out and furthermore we don't really need to provide a stable error interface that's disjoint from the protobuf.

A lot of the design decisions (esp around errors in this project) have been made assuming a need for interface stability independent from the protobuf file (i.e. mirroring Model and Error and Metrics with language native classes, adding placeholder error codes for categories of errors in case the server side adds new errors and we do not wish to update the proto file just yet): this is probably unnecessary. The main justification for this is interface simplicity: the Python protobuf objects are a bit wonky/unergonomic to actually use.
So anyways, we're now leaking Errors in the client side. The server side can stay as is (it's nice to have the error types subclass Exception).