server: switch to using a dictionary to find already loaded models
note for the future: this is not exactly the most efficient thing since we're storing all the models in memory twice (we're doing it once by keeping their flat buffer format in memory and tensorflow is doing it another time, presumably). Right now this is just fine; in the future it might become a concern.